+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant.cs
@@ -5,14 +5,21 @@ public class DeleteRestaurantCommand(int id) : ICommand
     public int Id { get; } = id;
 }
 
-public class DeleteRestaurantCommandHandler(IRestaurantsRepository repository, ILogger<DeleteRestaurantCommandHandler> logger)
+public class DeleteRestaurantCommandHandler(
+    IRestaurantsRepository repository,
+    IRestaurantAuthorizationService authService,
+    ILogger<DeleteRestaurantCommandHandler> logger)
     : ICommandHandler<DeleteRestaurantCommand>
 {
     public async Task<Unit> Handle(DeleteRestaurantCommand command, CancellationToken cancellationToken)
     {
         logger.LogInformation("Deleting restaurant with id: {RestaurantId}", command.Id);
 
-        var restaurant = await repository.GetByIdAsync(command.Id) ?? throw new NotFoundException(nameof(Restaurant), command.Id.ToString());
+        var restaurant = await repository.GetByIdAsync(command.Id) 
+            ?? throw new NotFoundException(nameof(Restaurant), command.Id.ToString());
+
+        if (!authService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbidException();
 
         await repository.DeleteAsync(restaurant);
